+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/NavBar.scala
@@ -7,7 +7,7 @@ import edu.gemini.seqexec.web.client.model.Pages.Root
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.vdom.html_<^._
 import edu.gemini.seqexec.web.client.OcsBuildInfo
-import edu.gemini.seqexec.web.client.semanticui.SemanticUI._
+import edu.gemini.web.client.facades.semanticui.SemanticUI._
 import edu.gemini.seqexec.web.client.semanticui.Size
 import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon._
 import edu.gemini.seqexec.web.client.semanticui.elements.menu.HeaderItem
@@ -76,7 +76,7 @@ object ConnectionState {
 
   case class Props(u: WebSocketConnection)
 
-  def formatTime(delay: Long): String = if (delay < 1000) {
+  def formatTime(delay: Int): String = if (delay < 1000) {
     f"${delay / 1000.0}%.1f"
   } else {
     f"${delay / 1000}%d"
